+++ inst/include/dplyr/MatrixColumnVisitor.h
@@ -142,6 +142,10 @@ namespace dplyr {
             return true ;
         }
         
+        bool is_na( int i ) const {
+            return false ;    
+        }
+        
     private:
         
         template <typename Container>