+++ tests/testthat/test-slice.r
@@ -75,3 +75,17 @@ test_that("slicing data table preserves input class", {
   expect_is(mtcars_dt %>% tbl_dt() %>% slice(1), "tbl_dt")
   expect_is(mtcars_dt %>% group_by(cyl) %>% slice(1), "grouped_dt")
 })
+
+test_that( "slice handles NA (#1235)", {
+  df <- data_frame( x = 1:3 )
+  expect_equal( nrow(slice(df, NA_integer_)), 0L )
+  expect_equal( nrow(slice(df, c(1L, NA_integer_))), 1L )
+  expect_equal( nrow(slice(df, c(-1L, NA_integer_))), 2L )
+  
+  df <- data_frame( x = 1:4, g = rep(1:2, 2) ) %>% group_by(g)
+  expect_equal( nrow(slice(df, NA)), 0L )
+  expect_equal( nrow(slice(df, c(1,NA))), 2 )
+  expect_equal( nrow(slice(df, c(-1,NA))), 2 )
+  
+})
+