+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * `mutate` and `arrange` works on empty data frames (#1142). 
 
+* `slice` handles NA (#1235). 
+
 # dplyr 0.4.2
 
 This is a minor release containing fixes for a number of crashes and issues identified by R CMD CHECK. There is one new "feature": dplyr no longer complains about unrecognised attributes, and instead just copies them over to the output.